Implement the initial project.
This service provides an HTTP server that allows users to
define JavaScript programs and later execute them with dynamic
parameters.

It's implemented using the Hyper HTTP library, which in turn is based on
Tokio.

A dedicated thread pool is setup to do JS execution.

Most settings are tuneable by specifying environment variables.